displayed the charts and added to the README
