+++ README.md
@@ -1,0 +1,26 @@
+# Expense Tracker
+
+## User Story
+As a user, I want to be able to track my expenses and sort them by category so I know where I am overspending and how much I want to be saving.
+
+## Description
+This expense tracker allows you to add any income you earn, and any expenses you make, and track them by category. It will then graph the categories to show what proportion of your income you spend in each area, e.g. bills, holidays. 
+
+## Usage
+To use this expense tracker, add all expenses with a description in the 'purchase' section. Any income can be added with a +£, and any expenses with a -£. 
+
+## Technologies Used
+For this project we used react, node, chart.js, re-charts, context. 
+
+## Future Development
+In the future, we will add a 'holiday' section, where the user can budget for their holidays and can use a currency converter and a Purchasing Power Parity API to see how far their budget would stretch in different currencies. 
+
+## Installation
+N/A
+
+## Credits
+
+## License
+Please refer to the MIT license in the repository. 
+
+## Badges
