+++ src/App-test.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import './App.css'
 import CategoryChart from './components/category-chart/categoryChart'
-import Balance from './components/balance indicator/balance'
+import Balance from './components/balance indicator/balanceIndicator'
 
 
 const App = () => {
