+++ src/App.jsx
@@ -5,6 +5,8 @@ import { IncomeExpenses } from './components/container/IncomeExpenses.jsx';
 import { TransactionList } from './components/container/TransactionList.jsx';
 import { AddTransaction } from './components/container/AddTransaction.jsx';
 import { GlobalProvider } from './context/GlobalState.jsx';
+import CategoryChart from './components/category-chart/categoryChart'
+import BalanceIndicator from './components/balance indicator/balanceIndicator.jsx'
 import 'bootstrap/dist/css/bootstrap.min.css';
 
 const App = () => {
@@ -14,8 +16,10 @@ const App = () => {
 <div>
   <Balance/>
 <IncomeExpenses/>
+<BalanceIndicator/>
 <TransactionList/>
 <AddTransaction/>
+<CategoryChart/>
 </div>
 </GlobalProvider>
