+++ src/components/category-chart/categoryChart.css
@@ -1,0 +1,5 @@
+.category-chart-container {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
