+++ src/main.jsx
@@ -1,8 +1,8 @@
 
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-// import App from './App.jsx'
-import App from './App-test.jsx'
+import App from './App.jsx'
+// import App from './App-test.jsx'
 import './index.css'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
